remove block of code that doesn't appear to be needed.
matches[3] will always be the same as the input text.  This code block appears to assume that the input is a valid JSON string, and looks like it is trying to replace the root object's attribute names with their values (if they attribute name has the specified prefix).  This doesn't seem correct or useful.